28-02 manha

+++ ocp/src/com/supportcom/ocp/service/PortalServiceImpl.java
@@ -3,10 +3,8 @@ package com.supportcom.ocp.service;
 import java.util.List;
 
 import com.supportcom.ocp.dao.PortalDAOImpl;
-import com.supportcomm.ocp.entity.Ivr;
 import com.supportcomm.ocp.entity.Portal;
 import com.supportcomm.ocp.entity.PortalLoginVO;
-import com.supportcomm.ocp.entity.Site;
 
 public class PortalServiceImpl implements PortalService {
 
@@ -44,17 +42,23 @@ public class PortalServiceImpl implements PortalService {
 	}
 
 	@Override
-	public void setSelectedAccessRights(List<Portal> selectedAccessRights) {
-		String[] reg =selectedAccessRights.get(0).getName().split(",");
-    	Long loginId= Long.parseLong( reg[1].toString());
-    	Long portalId= Long.parseLong( reg[0].toString());
-    	portalDao.removeAllPortalsLogin(loginId);
-    	for( Portal portal :selectedAccessRights){
-    		
-    		portalDao.associatePortalLogin(portalId, loginId);
-    		
-    	}
-
+	public List<Portal> setSelectedAccessRights(List<Portal> selectedAccessRights) {
+		Long loginId= 0l;
+		if (!selectedAccessRights.isEmpty()){
+			String[] reg =selectedAccessRights.get(0).getName().split(",");
+	    	loginId= Long.parseLong( reg[1].toString());
+	    	Long portalId= Long.parseLong( reg[0].toString());
+	    	portalDao.removeAllPortalsLogin(loginId);
+	    	for( Portal portal :selectedAccessRights){
+	    		
+	    		portalDao.associatePortalLogin(portalId, loginId);
+	    		
+	    	}
+		}
+    	
+    	List<PortalLoginVO> retorno = listAllByLogin( loginId);
+    	
+    	return  null;
 	}
 
 	public List<PortalLoginVO> listAllByLogin(Long loginId) {
